requirements.txt
Pillow~=8.4.0
datefinder~=0.7.1